refactor: restructure supabase and add category domain
- Migrate Supabase client and all repository imports to the standardized /src/infrastructure/repositories/supabase directory
- Update all existing service, hook, and test imports to use the new file paths
- Replace legacy single-file Category interface with proper domain entity, props interface, and CategoryType enum
- Add CategoryMapper to handle domain-persistence layer conversion
- Implement full CRUD TransactionRepository with test coverage
- Add initial category repository structure
- Remove unused TransactionType import from TransactionMapper
- Clean up and remove deprecated single-file Category entity
