--- src/domain/entities/account.ts
@@ -1,8 +1,0 @@
-export interface Account {
-  id: string;
-  name: string;
-  icon: string;
-  color: string;
-  user_id: string;
-  created_at: string;
-}
