+++ src/domain/entities/account/account.ts
@@ -1,0 +1,8 @@
+export interface Account {
+  id: string;
+  name: string;
+  icon: string;
+  color: string;
+  user_id: string;
+  created_at: string;
+}
