--- src/domain/entities/category.ts
@@ -1,11 +1,0 @@
-import { TransactionType } from '../enum/transaction-type';
-
-export interface Category {
-  id: string;
-  name: string;
-  icon: string;
-  color: string;
-  type: TransactionType;
-  user_id: string;
-  created_at: string;
-}
