+++ src/domain/enum/category-types.ts
@@ -1,0 +1,4 @@
+export enum CategoryType {
+  EXPENSE = 'EXPENSE',
+  INCOME = 'INCOME',
+}
